re-organize peak detection
