+++ src/spikeinterface/sortingcomponents/matching/method_list.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from .naive import NaiveMatching
-from .tdc import TridesclousPeeler
+from .tdc_peeler import TridesclousPeeler
 from .circus import CircusPeeler, CircusOMPSVDPeeler
 from .wobble import WobbleMatch
