+++ src/spikeinterface/sortingcomponents/peak_detection/__init__.py
@@ -1,0 +1,3 @@
+# from .method_list import matching_methods
+
+from .main import detect_peaks
