+++ src/spikeinterface/sortingcomponents/peak_detection/method_list.py
@@ -1,0 +1,15 @@
+from __future__ import annotations
+
+from .matched_filtering import MatchedFilteringPeakDetector
+from .by_channel import ByChannelPeakDetector, ByChannelTorchPeakDetector
+from .locally_exclusive import LocallyExclusivePeakDetector, LocallyExclusiveTorchPeakDetector, LocallyExclusiveOpenCLPeakDetector
+
+detection_methods = {
+    "locally_exclusive": LocallyExclusivePeakDetector,
+    "locally_exclusive_torch": LocallyExclusiveTorchPeakDetector,
+    "locally_exclusive_cl": LocallyExclusiveOpenCLPeakDetector,
+    "matched_filtering": MatchedFilteringPeakDetector,
+    "by_channel": ByChannelPeakDetector,
+    "by_channel_torch": ByChannelTorchPeakDetector
+}
+
