+++ src/spikeinterface/sortingcomponents/peak_detection/tests/test_peak_detection.py
@@ -10,16 +10,15 @@ from spikeinterface.sortingcomponents.peak_detection import detect_peaks
 
 from spikeinterface.core.node_pipeline import ExtractDenseWaveforms, ExtractSparseWaveforms
 from spikeinterface.sortingcomponents.peak_localization import LocalizeCenterOfMass
-from spikeinterface.sortingcomponents.features_from_peaks import PeakToPeakFeature
+from spikeinterface.sortingcomponents.waveforms.features_from_peaks import PeakToPeakFeature
 
 from spikeinterface.sortingcomponents.waveforms.temporal_pca import TemporalPCADenoising
-from spikeinterface.sortingcomponents.peak_detection import IterativePeakDetector
-from spikeinterface.sortingcomponents.peak_detection import (
-    DetectPeakByChannel,
-    DetectPeakByChannelTorch,
-    DetectPeakLocallyExclusive,
-    DetectPeakLocallyExclusiveTorch,
-)
+from spikeinterface.sortingcomponents.peak_detection.iterative import IterativePeakDetector
+
+
+from spikeinterface.sortingcomponents.peak_detection.by_channel import ByChannelPeakDetector, ByChannelTorchPeakDetector
+from spikeinterface.sortingcomponents.peak_detection.locally_exclusive import LocallyExclusivePeakDetector, LocallyExclusiveTorchPeakDetector
+
 
 from spikeinterface.core.node_pipeline import run_node_pipeline
 from spikeinterface.sortingcomponents.tools import get_prototype_and_waveforms_from_peaks
@@ -128,7 +127,7 @@ def peak_detector_kwargs_fixture(recording):
 
 
 def test_iterative_peak_detection(recording, job_kwargs, pca_model_folder_path, peak_detector_kwargs):
-    peak_detector_node = DetectPeakLocallyExclusive(**peak_detector_kwargs)
+    peak_detector_node = LocallyExclusivePeakDetector(**peak_detector_kwargs)
 
     ms_before = 1.0
     ms_after = 1.0
@@ -167,7 +166,7 @@ def test_iterative_peak_detection(recording, job_kwargs, pca_model_folder_path, 
 
 
 def test_iterative_peak_detection_sparse(recording, job_kwargs, pca_model_folder_path, peak_detector_kwargs):
-    peak_detector_node = DetectPeakLocallyExclusive(**peak_detector_kwargs)
+    peak_detector_node = LocallyExclusivePeakDetector(**peak_detector_kwargs)
 
     ms_before = 1.0
     ms_after = 1.0
@@ -212,7 +211,7 @@ def test_iterative_peak_detection_sparse(recording, job_kwargs, pca_model_folder
 
 
 def test_iterative_peak_detection_thresholds(recording, job_kwargs, pca_model_folder_path, peak_detector_kwargs):
-    peak_detector_node = DetectPeakLocallyExclusive(**peak_detector_kwargs)
+    peak_detector_node = LocallyExclusivePeakDetector(**peak_detector_kwargs)
 
     ms_before = 1.0
     ms_after = 1.0
@@ -366,10 +365,10 @@ def test_detect_peaks_locally_exclusive_matched_filtering(recording, job_kwargs)
 
 
 detection_classes = [
-    DetectPeakByChannel,
-    DetectPeakByChannelTorch,
-    DetectPeakLocallyExclusive,
-    DetectPeakLocallyExclusiveTorch,
+    ByChannelPeakDetector,
+    ByChannelTorchPeakDetector,
+    LocallyExclusivePeakDetector,
+    LocallyExclusiveTorchPeakDetector,
 ]
 
 
@@ -390,7 +389,7 @@ def test_peak_sign_consistency(recording, job_kwargs, detection_class):
     # To account for exclusion of positive peaks that are to close to negative peaks.
     # This should be excluded by the detection method when is exclusive so using peak_sign="both" should
     # Generate less peaks in this case
-    if detection_class not in (DetectPeakByChannelTorch, DetectPeakLocallyExclusiveTorch):
+    if detection_class not in (ByChannelTorchPeakDetector, LocallyExclusiveTorchPeakDetector):
         # TODO later Torch do not pass this test
         assert (negative_peaks.size + positive_peaks.size) >= all_peaks.size
 
@@ -534,7 +533,7 @@ if __name__ == "__main__":
     #     recording, job_kwargs_main, pca_model_folder_path_main, peak_detector_kwargs_main
     # )
 
-    # test_peak_sign_consistency(recording, torch_job_kwargs_main, DetectPeakLocallyExclusiveTorch)
+    # test_peak_sign_consistency(recording, torch_job_kwargs_main, LocallyExclusiveTorchPeakDetector)
     # test_peak_detection_with_pipeline(recording, job_kwargs_main, torch_job_kwargs_main, tmp_path)
 
     test_detect_peaks_locally_exclusive_matched_filtering(
